XCOFF: Ignore hidden linkage for now
This ought to be treated the same as LGlobal, but both MCSA_Global and
MCSA_Hidden can show up on the same symbol, and that results in an
error for redefining the storage class. Since we're only going to
statically link objects, it doesn't really make a difference whether
we get hidden or public symbols, except in forgoing some dead code
stripping from the linker.